dfir_lang::graph

Struct AsCodeOptions

#[non_exhaustive]
pub struct AsCodeOptions {
    pub exclude_type_guards: bool,
    pub exclude_meta: bool,
    pub include_metrics_tracking: bool,
}

Options for DfirGraph::as_code_with_options.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

exclude_type_guards: bool

Controls whether type guards are emitted in codegen. Excluding type guards may result in worse error messages, and may have little benefit on --release builds.

exclude_meta: bool

Controls whether the runtime meta graph + diagnostics JSON blobs are baked into the generated Dfir::new(...) call.

include_metrics_tracking: bool

Controls whether metrics are tracked. Metrics tracking is opt-in: no tracking code is generated unless this is set. Even if metrics are tracked, they still need to be reported via the Context::metrics field.