dfir_lang::graph

Struct FlatGraphBuilder

pub struct FlatGraphBuilder { /* private fields */ }

Wraper around DfirGraph to build a flat graph from AST code.

Implementations

impl FlatGraphBuilder

pub fn new() -> Self

Create a new empty graph builder.

pub fn from_dfir(input: DfirCode) -> Self

Convert the DFIR code AST into a graph builder.

pub fn build(self) -> Result<FlatGraphBuilderOutput, Diagnostics>

Build into an unpartitioned DfirGraph, returning a struct containing the flat graph, any diagnostics, and other outputs.

If any diagnostics are errors, Err is returned and the underlying graph is lost.

pub fn add_dfir( &mut self, dfir: DfirCode, current_loop: Option<GraphLoopId>, operator_tag: Option<&str>, )

Adds all DfirStatements within the DfirCode to this DfirGraph.

Optional configuration:

In the given loop context current_loop.

With the given operator tag operator_tag.

pub fn add_statement(&mut self, stmt: DfirStatement)

Add a single DfirStatement line to this DfirGraph in the root context.

pub fn append_assign_pipeline( &mut self, asgn_name: Option<&Ident>, pred_name: Option<&Ident>, pipeline: Pipeline, current_loop: Option<GraphLoopId>, operator_tag: Option<&str>, )

Programatically add an pipeline, optionally adding pred_name as a single predecessor and assigning it all to asgn_name.

In DFIR syntax, equivalent to Self::add_statement of (if all names are supplied):

#asgn_name = #pred_name -> #pipeline;

But with, optionally:

A current_loop to put the operator in.

An operator_tag to tag the operator with, for debugging/tracing.